quantumgraph/qgis_plugin_tools/requirements-dev.txt
# Testing
pytest-qgis~=1.2.0
pytest-timeout==2.0.2
pytest-qt==3.3.0
pytest-mock==3.6.1

requests==2.26.0
types-requests==2.27.7
